% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{additional_aliases_modal}
\alias{additional_aliases_modal}
\title{Show a pop-up modal if the uploaded panel has additional gene aliases}
\usage{
additional_aliases_modal()
}
\description{
Show a pop-up modal if the uploaded panel has additional gene aliases
}
