% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{additional_aliases_model}
\alias{additional_aliases_model}
\title{Show a pop-up modal if the uploaded panel has additional gene aliases}
\usage{
additional_aliases_model(original_gene_list, alias_list)
}
\arguments{
\item{original_gene_list}{The original gene list uploaded}

\item{merged_list}{The list with the original genes as well as the alternative aliases}
}
\description{
Show a pop-up modal if the uploaded panel has additional gene aliases
}
