% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{assay_modal}
\alias{assay_modal}
\title{Show an input modal to select the single cell assay to use for analysis}
\usage{
assay_modal(assays, failed = FALSE)
}
\description{
Show an input modal to select the single cell assay to use for analysis
}
