% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{cell_type_ignored_modal}
\alias{cell_type_ignored_modal}
\title{Show a shinyalert warning if cell types retained for analysis are below the minimum cell cutoff.
cytomarker will ignore these for analysis and print their names during analysis.}
\usage{
cell_type_ignored_modal(min_cat_count, excluded_cells)
}
\arguments{
\item{min_cat_count}{The current value of minimum cell cutoff (minimum of 2).}

\item{the}{vector of the names of cell types below the cutoff to be ignored.}
}
\description{
Show a shinyalert warning if cell types retained for analysis are below the minimum cell cutoff.
cytomarker will ignore these for analysis and print their names during analysis.
}
