% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{check_for_human_genes}
\alias{check_for_human_genes}
\title{Check if the majority of genes are not human in an SCE}
\usage{
check_for_human_genes(sce)
}
\arguments{
\item{sce}{The SingleCellExperiment object}
}
\description{
Check if the majority of genes are not human in an SCE
}
