% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{check_rowData_for_hugo}
\alias{check_rowData_for_hugo}
\title{Checks if there are any columns in rowData that contain gene names}
\usage{
check_rowData_for_hugo(sce, grch38)
}
\description{
Checks if there are any columns in rowData that contain gene names
}
