% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{compute_fm}
\alias{compute_fm}
\title{Compute the findMarkers outputs and store}
\usage{
compute_fm(
  sce,
  columns,
  pref_assay,
  allowed_genes,
  p_val_type = "all",
  use_rank = T
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{columns}{A vector storing columns}

\item{pref_assay}{Assay loaded}

\item{p_val_type}{The type of p-value to use in marker group calculation. Can be either all or any.}

\item{use_rank}{Whether or not to use the logFC rank as the sorting criteria. Default is TRUE}
}
\description{
Note: this used to be part of get_markers but needs
split off after colouring
}
