% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_column_to_character_or_factor}
\alias{convert_column_to_character_or_factor}
\title{Convert a SingleCellExperiment column to a factor or a character}
\usage{
convert_column_to_character_or_factor(sce, input_column)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{input_column}{a metadata column in the sce input}
}
\description{
Convert a SingleCellExperiment column to a factor or a character
}
