% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_global_colour_palette}
\alias{create_global_colour_palette}
\title{Create the global cytosel palette with or without seeding. Note: repeats unique colours twice as the palette
generates only 74 unique colours, for a total palette length of 148.}
\usage{
create_global_colour_palette(pal_seed = NULL)
}
\arguments{
\item{pal_seed}{random seed used to shuffle the palette}
}
\description{
Create the global cytosel palette with or without seeding. Note: repeats unique colours twice as the palette
generates only 74 unique colours, for a total palette length of 148.
}
