% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_heatmap}
\alias{create_heatmap}
\title{Make the expression and correlation heatmaps}
\usage{
create_heatmap(
  sce,
  markers,
  column,
  display,
  normalization,
  pref_assay = "logcounts"
)
}
\description{
Make the expression and correlation heatmaps
}
