% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_keep_vector_during_subsetting}
\alias{create_keep_vector_during_subsetting}
\title{Given a vector of membership during subsetting, create a sce column for retaining during analysis}
\usage{
create_keep_vector_during_subsetting(sce, vec_to_keep)
}
\description{
Given a vector of membership during subsetting, create a sce column for retaining during analysis
}
