% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{create_multimarker_frame}
\alias{create_multimarker_frame}
\title{Collect possible multimarkers from the output of marker finding}
\usage{
create_multimarker_frame(frame, recommended_markers)
}
\arguments{
\item{frame}{A dataframe containing the markers selected in \code{findMarkers}}

\item{recommended_markers}{A vector of the currently recommended markers for a panel}
}
\description{
Collect possible multimarkers from the output of marker finding
}
