% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_table_of_hetero_cat}
\alias{create_table_of_hetero_cat}
\title{Create a data frame table of counts for a hetergeneity category}
\usage{
create_table_of_hetero_cat(sce, metadata_column)
}
\description{
Create a data frame table of counts for a hetergeneity category
}
