% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{curated_dataset_modal}
\alias{curated_dataset_modal}
\title{Show an input modal for the user to select a pre-curated cytosel dataset}
\usage{
curated_dataset_modal()
}
\description{
Show an input modal for the user to select a pre-curated cytosel dataset
}
