% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{current_pan_not_valid_modal}
\alias{current_pan_not_valid_modal}
\title{Show an input modal if the current panel contains genes that are not found in the newest uploaded SCE}
\usage{
current_pan_not_valid_modal(missing_genes)
}
\arguments{
\item{missing_genes}{vector of genes that are not in the current SCE but in the current marker panel}
}
\description{
Show an input modal if the current panel contains genes that are not found in the newest uploaded SCE
}
