% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{detect_assay_and_create_logcounts}
\alias{detect_assay_and_create_logcounts}
\title{This function detects whether the logcounts assay are indeed logcounts by finding
the residual of the logcount expression assays' rowsums divided by one. If this
is zero they must be counts and are converted to logcounts}
\usage{
detect_assay_and_create_logcounts(sce)
}
\arguments{
\item{sce}{Uploaded SingleCellExperiment}
}
\description{
This function detects whether the logcounts assay are indeed logcounts by finding
the residual of the logcount expression assays' rowsums divided by one. If this
is zero they must be counts and are converted to logcounts
}
