% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{detect_umap_dims_in_sce}
\alias{detect_umap_dims_in_sce}
\title{Detect if the sce has any UMAP dimension assays}
\usage{
detect_umap_dims_in_sce(sce)
}
\arguments{
\item{sce}{scRNAseq dataset}
}
\description{
Detect if the sce has any UMAP dimension assays
}
