% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{dne_modal}
\alias{dne_modal}
\title{Show a shinyalert error if the input marker is not found in the current antibody set}
\usage{
dne_modal(dne)
}
\arguments{
\item{dne}{the string name of the selected marker}
}
\description{
Show a shinyalert error if the input marker is not found in the current antibody set
}
