% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_allowed_genes}
\alias{get_allowed_genes}
\title{Picks out allowed genes corresponding
to antibody applications}
\usage{
get_allowed_genes(selected_applications, applications_parsed, sce)
}
\description{
Picks out allowed genes corresponding
to antibody applications
}
