% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_antibody_applications}
\alias{get_antibody_applications}
\title{Parse out antibody applications}
\usage{
get_antibody_applications(
  antibody_info,
  column_gene = "Symbol",
  column_application = "Listed Applications"
)
}
\description{
TODO: when the antibody application list is fixed, this can
be parsed ahead of time
}
