% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_antibody_info}
\alias{get_antibody_info}
\title{Get antibody info/status for a particular gene}
\usage{
get_antibody_info(gene_id, df)
}
\arguments{
\item{gene_id}{An element in the marker list}

\item{df}{A dataframe storing antibody information}
}
\description{
Get antibody info/status for a particular gene
}
