% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gene_aliases}
\alias{get_gene_aliases}
\title{Get a list of aliases for the marker list (all possible gene IDs that match to an Entrez ID)}
\usage{
get_gene_aliases(marker_list, gene_mapper, allowed_genes)
}
\arguments{
\item{marker_list}{A vector of markers in gene symbol format}

\item{gene_mapper}{A data frame that maps gene symbols to Entrez IDs}

\item{allowed_genes}{Vector of genes permitted in analysis}
}
\description{
Get a list of aliases for the marker list (all possible gene IDs that match to an Entrez ID)
}
