% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_legend}
\alias{get_legend}
\title{Return the legend of the cell type colours}
\usage{
get_legend(palette)
}
\arguments{
\item{palette}{A vector storing hex colours}
}
\description{
Return the legend of the cell type colours
}
