% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_markers}
\alias{get_markers}
\title{Compute markers}
\usage{
get_markers(fms, panel_size)
}
\arguments{
\item{fms}{Stored findMarkers outputs}

\item{panel_size}{Targeted number of markers in selected panel}
}
\description{
Compute markers
}
