% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_scores}
\alias{get_scores}
\title{Compute the metric scores}
\usage{
get_scores(sce, columns, mrkrs, max_cells = 5000, pref_assay = "logcounts")
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{columns}{A vector storing columns}

\item{mrkrs}{A list storing three lists of markers:
recommended_markers, scratch_markers, and top_markers}

\item{max_cells}{Maximum number of cells}

\item{pref_assay}{Assay loaded, default is 'logcounts'}
}
\description{
Compute the metric scores
}
