% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_scores_one_column}
\alias{get_scores_one_column}
\title{Compute the metric scores for a specific column}
\usage{
get_scores_one_column(
  sce_tr,
  column,
  mrkrs,
  max_cells = 5000,
  pref_assay = "logcounts"
)
}
\arguments{
\item{sce_tr}{A SingleCellExperiment object}

\item{column}{A vector storing a column}

\item{mrkrs}{A list storing three lists of markers:
recommended_markers, scratch_markers, and top_markers}

\item{max_cells}{Maximum number of cells}

\item{pref_assay}{Assay loaded, default is 'logcounts'}
}
\description{
Compute the metric scores for a specific column
}
