% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltips.R
\name{get_tooltip_q_coldata_column}
\alias{get_tooltip_q_coldata_column}
\title{Get the tooltip info for cell category input}
\usage{
get_tooltip_q_coldata_column()
}
\description{
Get the tooltip info for cell category input
}
