% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_umap}
\alias{get_umap}
\title{Compute the UMAP}
\usage{
get_umap(
  sce,
  columns,
  pref_assay,
  precomputed_vals,
  dim_col,
  only_top_markers = F
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{columns}{A vector storing columns}

\item{pref_assay}{Assay loaded}

\item{precomputed_vals}{A boolean stating whether or not the given sce has a precomputed UMAP dim}

\item{dim_col}{If the precomputed UMAP values exist, what is the name of the dimension in the sce}

\item{only_top_markers}{Boolean stating whether or not the UMAP should use only top markers (Default is False)}
}
\description{
Compute the UMAP
}
