% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{good_col}
\alias{good_col}
\title{Split columns into usable (good) and unusable (bad) categories
to be passed on to compute_fm}
\usage{
good_col(sce, column)
}
\description{
Split columns into usable (good) and unusable (bad) categories
to be passed on to compute_fm
}
