% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{high_cell_number_warning}
\alias{high_cell_number_warning}
\title{Show a shinyalert warning if the sce for analysis has more than a set threshold of cell counts}
\usage{
high_cell_number_warning(ncell_sce, num_cells)
}
\arguments{
\item{ncell_sce}{The number of cells in the Singlecellexperiment object used for analysis}

\item{num_cells}{The threshold for cell number for the sce. An ace with a cell number greater than this threshold will trigger the warning.}
}
\description{
Show a shinyalert warning if the sce for analysis has more than a set threshold of cell counts
}
