% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_row}
\alias{in_row}
\title{#' replace a null, NA, or empty value with a string representation "None
#' @param val the list element to convert
replace_na_null_empty <- function(val) {
if (is_empty(val)) {
return ("None")
} else if (is.na(val)) {
return ("None")
} else {
return(val)
}
}
Filter out antibody table rows by a vector list of antibody applications}
\usage{
in_row(applications, row)
}
\arguments{
\item{applications}{A vector of antibody applications}

\item{row}{The string in a row representing the applications for a specific gene in the table}
}
\description{
#' replace a null, NA, or empty value with a string representation "None
#' @param val the list element to convert
replace_na_null_empty <- function(val) {
if (is_empty(val)) {
return ("None")
} else if (is.na(val)) {
return ("None")
} else {
return(val)
}
}
Filter out antibody table rows by a vector list of antibody applications
}
