% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{invalid_metadata_modal}
\alias{invalid_metadata_modal}
\title{Show a shinyalert error if the metadata column selected has too many levels}
\usage{
invalid_metadata_modal(column)
}
\arguments{
\item{column}{the string name of the metadata column selected}
}
\description{
Show a shinyalert error if the metadata column selected has too many levels
}
