% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_violin_plot}
\alias{make_violin_plot}
\title{Create a violin plot from a vector of genes, a cell type category in an SCE, and the assay desired}
\usage{
make_violin_plot(sce, genes, cell_type, assay)
}
\arguments{
\item{sce}{the SingleCellExperiment object}

\item{genes}{The vector of genes to visualize}

\item{cell_type}{The cell type category to resolve using the gene vector}

\item{assay}{The name of the assay to use. Common selections are logcounts or counts}
}
\description{
Create a violin plot from a vector of genes, a cell type category in an SCE, and the assay desired
}
