% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_gene_name_to_antibody_icon}
\alias{map_gene_name_to_antibody_icon}
\title{Map gene symbols to antibody icons depending on
whether the gene is found in the database}
\usage{
map_gene_name_to_antibody_icon(gene_id, df_antibody)
}
\description{
Map gene symbols to antibody icons depending on
whether the gene is found in the database
}
