% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{no_cells_left_modal}
\alias{no_cells_left_modal}
\title{Show a shinyalert error if, user subsetting an filtering, no cells remain for analysis}
\usage{
no_cells_left_modal()
}
\description{
Show a shinyalert error if, user subsetting an filtering, no cells remain for analysis
}
