% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{panel_too_small_modal}
\alias{panel_too_small_modal}
\title{Show an error modal if the panel size is less than 2 (too small)}
\usage{
panel_too_small_modal(size)
}
\arguments{
\item{size}{The integer value of the requested panel size}
}
\description{
Show an error modal if the panel size is less than 2 (too small)
}
