% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{parse_gene_names}
\alias{parse_gene_names}
\title{Parses the gene names in a sce object
This function tries to find gene symbols by going through several steps:
\enumerate{
\item Count the number and proportion of rownames that are listed in the
symbol column of the annotables grch38 object.
If the proportion is more than 50\% and the symbols are human they are used
If the proportion is less than 50\%, more than 100 genes are found and
the symbols are human they are still used
\item Check if there are any human gene names in \code{rowData(sce)} and if more
than 50\%/100 genes are in the symbol column of the annotables grch38 object
\item If rownames are not NULL, check if they are ensembl ID's and convert to symbol
}}
\usage{
parse_gene_names(sce, grch38, check_genes = T)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{grch38}{Dataframe with two columns: \code{ensgene} and \code{symbol}, (originally from the annotables package)}

\item{check_genes}{If TRUE, checks the genes against human genes. Otherwise, uses the gene names found in the dataset
analyses are removed. The following genes are removed: ribosomal proteins, mitochondrial ribosomal
and other mitochondrial proteins, heat shock proteins, Jun, Fos and Malat1}

\item{remove_confounding_genes}{If true (default) genes that frequently confound single cell}
}
\value{
genes A vector of gene names
}
\description{
Parses the gene names in a sce object
This function tries to find gene symbols by going through several steps:
\enumerate{
\item Count the number and proportion of rownames that are listed in the
symbol column of the annotables grch38 object.
If the proportion is more than 50\% and the symbols are human they are used
If the proportion is less than 50\%, more than 100 genes are found and
the symbols are human they are still used
\item Check if there are any human gene names in \code{rowData(sce)} and if more
than 50\%/100 genes are in the symbol column of the annotables grch38 object
\item If rownames are not NULL, check if they are ensembl ID's and convert to symbol
}
}
