% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{pre_computed_umap_modal}
\alias{pre_computed_umap_modal}
\title{Show an input modal if the user wants to use precomputed UMAP coordinates}
\usage{
pre_computed_umap_modal(possible_dims)
}
\description{
Show an input modal if the user wants to use precomputed UMAP coordinates
}
