% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_back_in_saved_yaml}
\alias{read_back_in_saved_yaml}
\title{read back in a saved config yaml to resume analysis}
\usage{
read_back_in_saved_yaml(yaml)
}
\arguments{
\item{yaml}{Path to a yaml file with a saved configuration}
}
\description{
read back in a saved config yaml to resume analysis
}
