% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_cell_types_by_min_counts}
\alias{remove_cell_types_by_min_counts}
\title{Create a vector of metadata values that pass a minimum count}
\usage{
remove_cell_types_by_min_counts(
  grouped_frame,
  sce,
  metadata_column,
  min_counts
)
}
\description{
Create a vector of metadata values that pass a minimum count
}
