% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{remove_confounding_genes}
\alias{remove_confounding_genes}
\title{Remove confounding genes in the dataset}
\usage{
remove_confounding_genes(gene_vector)
}
\arguments{
\item{gene_vector}{A vector of genes to clean}
}
\description{
Remove confounding genes in the dataset
}
