% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{remove_null_and_va_from_cell_cat}
\alias{remove_null_and_va_from_cell_cat}
\title{Update the keep_for_analysis vector in an SCE to exclude null and NA values}
\usage{
remove_null_and_va_from_cell_cat(sce, input_column)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{input_column}{The string name of the desired input column in sce}
}
\description{
Update the keep_for_analysis vector in an SCE to exclude null and NA values
}
