% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replace_missing_types}
\alias{replace_missing_types}
\title{Replace a zero character or NA with a new string}
\usage{
replace_missing_types(current_string, replacement_string)
}
\arguments{
\item{current_string}{The current zero character string}

\item{replacement_string}{The string to replace the zero character string}
}
\description{
Replace a zero character or NA with a new string
}
