% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_na_null_empty}
\alias{replace_na_null_empty}
\title{replace a null, NA, or empty value with a string representation "None}
\usage{
replace_na_null_empty(val)
}
\arguments{
\item{val}{the list element to convert}
}
\description{
replace a null, NA, or empty value with a string representation "None
}
