% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{reset_analysis_modal}
\alias{reset_analysis_modal}
\title{Show an input modal to confirm resetting the analysis}
\usage{
reset_analysis_modal()
}
\description{
Show an input modal to confirm resetting the analysis
}
