% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{reset_option_on_change_modal}
\alias{reset_option_on_change_modal}
\title{Show an input modal to give the user the option to reset the panel on}
\usage{
reset_option_on_change_modal(change)
}
\arguments{
\item{change}{The type of change detected in the configuration}
}
\description{
Show an input modal to give the user the option to reset the panel on
}
