% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{reupload_before_sce_modal}
\alias{reupload_before_sce_modal}
\title{Show a shinyalert error if the user tries to upload the yml file before the sce}
\usage{
reupload_before_sce_modal()
}
\description{
Show a shinyalert error if the user tries to upload the yml file before the sce
}
