% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{reupload_error_modal}
\alias{reupload_error_modal}
\title{Show a shinyalert error if something goes wrong with the re-upload from yml file}
\usage{
reupload_error_modal(message)
}
\arguments{
\item{message}{The specific message type to be displayed depending on the reu-upload error}
}
\description{
Show a shinyalert error if something goes wrong with the re-upload from yml file
}
