% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{reupload_failed_modal}
\alias{reupload_failed_modal}
\title{Show a shinyalert error if the reupload files with the selected sce}
\usage{
reupload_failed_modal()
}
\description{
Show a shinyalert error if the reupload files with the selected sce
}
