% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{reupload_warning_modal}
\alias{reupload_warning_modal}
\title{Show a shinyalert warning if the dimensions of the current sce do not match the input yaml}
\usage{
reupload_warning_modal(title_message, body_message)
}
\arguments{
\item{title_message}{The type of warning message in the title}

\item{body_message}{The type of warning message in the body}
}
\description{
Show a shinyalert warning if the dimensions of the current sce do not match the input yaml
}
