% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{set_current_markers_safely}
\alias{set_current_markers_safely}
\title{Controlled way of setting current markers
Return them in a safe way!}
\usage{
set_current_markers_safely(markers, fms, default_type = NULL)
}
\arguments{
\item{markers}{A list storing three lists of markers:
recommended_markers, scratch_markers, and top_markers}

\item{fms}{Stored findMarkers outputs}
}
\description{
Controlled way of setting current markers
Return them in a safe way!
}
