% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{set_text_colour_based_on_background}
\alias{set_text_colour_based_on_background}
\title{Convert a text colour into black or white based on the RGB values of its background to improve visibility}
\usage{
set_text_colour_based_on_background(text_colour)
}
\arguments{
\item{text_colour}{the colour to be converted into black or white}
}
\description{
Convert a text colour into black or white based on the RGB values of its background to improve visibility
}
