% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{show_alias_table_modal}
\alias{show_alias_table_modal}
\title{Show an input modal to view the table of gene aliases for the selected panel and dataset}
\usage{
show_alias_table_modal()
}
\description{
Show an input modal to view the table of gene aliases for the selected panel and dataset
}
