% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{subsampling_error_modal}
\alias{subsampling_error_modal}
\title{Show an error modal if, after subsampling, certain cell types do not meet the minimum count threshold}
\usage{
subsampling_error_modal(cell_types)
}
\arguments{
\item{cell_types}{A vector of the cell types that do not pass the count threshold}
}
\description{
Show an error modal if, after subsampling, certain cell types do not meet the minimum count threshold
}
