% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{suggestion_modal}
\alias{suggestion_modal}
\title{Show an input modal to select the markers to remove from the selected marker space to scratch}
\usage{
suggestion_modal(failed = FALSE, suggestions)
}
\description{
Show an input modal to select the markers to remove from the selected marker space to scratch
}
