% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{threshold_too_low_modal}
\alias{threshold_too_low_modal}
\title{Show a shinyalert error if the minimum cell cutoff input is set below 2}
\usage{
threshold_too_low_modal()
}
\description{
Show a shinyalert error if the minimum cell cutoff input is set below 2
}
