% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{throw_error_or_warning}
\alias{throw_error_or_warning}
\title{Global function to throw errors/warnings}
\usage{
throw_error_or_warning(
  type = "notification",
  message,
  duration = NULL,
  notificationType = "default",
  errorTitle = "Error"
)
}
\arguments{
\item{type}{needs to be \code{error} or \code{notification}. \code{error} will create a popup error
while \code{notification} will display a notification in the bottom right. Defaults to \code{notification}.}

\item{message}{the message to be displayed to the user}

\item{duration}{number of seconds to display message before it disappears.
Defaults to NULL (does not close until closed by user).}

\item{notificationType}{Controls the colour of the message. Allowed values are 'default' (grey),
'message' (blue), 'warning' (yellow) and 'error' (red). Only applicable to notifications}

\item{errorTitle}{Title to display on the error message. Defaults to 'Error'}
}
\description{
Global function to throw errors/warnings
}
