% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{time_zone_modal}
\alias{time_zone_modal}
\title{Allow the user to change the time zone in the modal}
\usage{
time_zone_modal(possible_time_zones, current_input)
}
\arguments{
\item{possible_time_zones}{Vector of possible global time zones}

\item{current_input}{Currently selected time zone}
}
\description{
Allow the user to change the time zone in the modal
}
