% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{too_large_to_show_modal}
\alias{too_large_to_show_modal}
\title{Show a shinyalert error if the user tries to render the distribution table with more than 100 unique elements}
\usage{
too_large_to_show_modal(col_name)
}
\arguments{
\item{col_name}{name of column in colData of current scRNAseq dataset}
}
\description{
Show a shinyalert error if the user tries to render the distribution table with more than 100 unique elements
}
