% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{unique_element_modal}
\alias{unique_element_modal}
\title{Show a shinyalert error if either 1. the selected column has only one type, or
2. the selected column has over 100 unique types}
\usage{
unique_element_modal(col)
}
\description{
Show a shinyalert error if either 1. the selected column has only one type, or
2. the selected column has over 100 unique types
}
