% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{warning_modal}
\alias{warning_modal}
\title{Show a shinyalert warning if the uploaded marker is not found in the sce used for analysis}
\usage{
warning_modal(not_sce)
}
\description{
Show a shinyalert warning if the uploaded marker is not found in the sce used for analysis
}
