% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_proportion_in_annotables}
\alias{calculate_proportion_in_annotables}
\title{Helper function for \code{parse_gene_names}
Calculates the overlap of gene vector with annotables}
\usage{
calculate_proportion_in_annotables(g, annotation)
}
\description{
Helper function for \code{parse_gene_names}
Calculates the overlap of gene vector with annotables
}
