% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{check_rownames_for_ensembl}
\alias{check_rownames_for_ensembl}
\title{Checks if the rownames of the sce contain human ensembl ID's}
\usage{
check_rownames_for_ensembl(sce, grch38)
}
\description{
Checks if the rownames of the sce contain human ensembl ID's
}
