% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_alternatives}
\alias{compute_alternatives}
\title{Compute alternative markers}
\usage{
compute_alternatives(gene_to_replace, sce, pref_assay, n_correlations)
}
\description{
Compute alternative markers
}
