% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_alternatives}
\alias{compute_alternatives}
\title{Compute alternative markers}
\usage{
compute_alternatives(
  gene_to_replace,
  sce,
  pref_assay = "logcounts",
  n_correlations
)
}
\arguments{
\item{gene_to_replace}{Gene to be replaced}

\item{sce}{A SingleCellExperiment object}

\item{pref_assay}{Assay loaded, default is 'logcounts'}

\item{n_correlations}{Number of markers to replace gene_to_replace}
}
\description{
Compute alternative markers
}
