% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_fm}
\alias{compute_fm}
\title{Compute the findMarkers outputs and store}
\usage{
compute_fm(sce, columns, pref_assay = "logcounts")
}
\description{
Note: this used to be part of get_markers but needs
split off after colouring
}
