% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_global_colour_palette}
\alias{create_global_colour_palette}
\title{Filter a singlecellexperiment for metadata types with low counts. NOTE: repeats unique colours as the palette
generates only 74 unique colours.}
\usage{
create_global_colour_palette(pal_seed = NULL)
}
\arguments{
\item{pal_seed}{random seed used to shuffle the palette}
}
\description{
Filter a singlecellexperiment for metadata types with low counts. NOTE: repeats unique colours as the palette
generates only 74 unique colours.
}
