% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{create_global_colour_palette}
\alias{create_global_colour_palette}
\title{Create the global cytomarker palette with or without seeding. The palette begins with 12 color-blind friendly colours
then moves into 74 uniquely generated colors form brewer.pal, ending with 2 repeats from the first vector for
a final vector of 100 unique colours..}
\usage{
create_global_colour_palette(pal_seed = NULL)
}
\arguments{
\item{pal_seed}{random seed used to shuffle the palette (Default is NULL)}
}
\description{
Create the global cytomarker palette with or without seeding. The palette begins with 12 color-blind friendly colours
then moves into 74 uniquely generated colors form brewer.pal, ending with 2 repeats from the first vector for
a final vector of 100 unique colours..
}
