% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_heatmap}
\alias{create_heatmap}
\title{Make the expression and correlation heatmaps}
\usage{
create_heatmap(sce, markers, column, display, normalization, pref_assay)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{markers}{A list storing three lists of markers:
recommended_markers, scratch_markers, and top_markers}

\item{column}{A vector storing a column}

\item{display}{A vector storing selected heatmap display option}

\item{normalization}{A vector holding either 'Expression' or 'z-score'}

\item{pref_assay}{Assay loaded}
}
\description{
Make the expression and correlation heatmaps
}
