% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{create_run_param_list}
\alias{create_run_param_list}
\title{create a list of the run parameters for download or run logging}
\usage{
create_run_param_list(
  marker_list,
  input_file,
  assay_used,
  het_source,
  panel_size,
  cell_cutoff_value,
  subsample,
  subsample_number,
  marker_strat,
  antibody_apps,
  selected_cell_types,
  precomputed_umap_used,
  num_cells,
  num_genes,
  metrics
)
}
\description{
create a list of the run parameters for download or run logging
}
