% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{curated_dataset_modal}
\alias{curated_dataset_modal}
\title{Show an input modal for the user to select a pre-curated cytosel dataset}
\usage{
curated_dataset_modal(compartment_options, dataset_options, failed = FALSE)
}
\arguments{
\item{compartment_options}{a vector of the possible cell type supersets as established by Tabula Sapiens}

\item{dataset_options}{a vector of the identifiers for the possible loadable datasets}
}
\description{
Show an input modal for the user to select a pre-curated cytosel dataset
}
