% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_data}
\alias{download_data}
\title{Download all marker data to a zip file}
\usage{
download_data(zip_filename, config, plots, heatmap, antibody_table)
}
\description{
Download all marker data to a zip file
}
