% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cell_type_add_markers_reactable}
\alias{get_cell_type_add_markers_reactable}
\title{Get the reactable for the modal dialogue when add
markers for a given cell type}
\usage{
get_cell_type_add_markers_reactable(fm, current_markers)
}
\description{
Get the reactable for the modal dialogue when add
markers for a given cell type
}
