% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_markers}
\alias{get_markers}
\title{Compute markers}
\usage{
get_markers(fms, panel_size, marker_strategy, sce)
}
\arguments{
\item{fms}{Stored findMarkers outputs}

\item{panel_size}{Targeted number of markers in selected panel}

\item{marker_strategy}{Strategy for picking markers, either heuristic
cell type based or geneBasis}

\item{sce}{SingleCellExperiment object}
}
\description{
Compute markers
}
