% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_umap}
\alias{get_umap}
\title{Compute the UMAP}
\usage{
get_umap(
  sce,
  columns,
  pref_assay,
  precomputed_vals = NULL,
  dim_col = NULL,
  only_top_markers = F
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{columns}{A vector storing columns}

\item{pref_assay}{Assay loaded}
}
\description{
Compute the UMAP
}
