% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{good_col}
\alias{good_col}
\title{Compute the number of unique elements in each column for compute_fm}
\usage{
good_col(sce, column)
}
\description{
Compute the number of unique elements in each column for compute_fm
}
