% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{good_col}
\alias{good_col}
\title{Split columns into usable (good) and unusable (bad) categories
to be passed on to compute_fm}
\usage{
good_col(sce, columns)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{columns}{A vector storing columns}
}
\description{
Split columns into usable (good) and unusable (bad) categories
to be passed on to compute_fm
}
