% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{map_gene_name_to_antibody_icon}
\alias{map_gene_name_to_antibody_icon}
\title{Return the legend of the cell type colours}
\usage{
map_gene_name_to_antibody_icon(gene_id, markers)
}
\arguments{
\item{gene_id}{An element in the marker list}

\item{markers}{the list of markers}
}
\description{
Map gene symbols to antibody icons depending on
whether the gene is initially suggested or selected by the user
}
