% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_input_scrnaseq}
\alias{read_input_scrnaseq}
\title{Read the input single-cell RNA-seq dataset
from compressed RDS file. Accepts either:
(1) SingleCellExperiment
(2) Seurat object (converted to SingleCellExperiment)}
\usage{
read_input_scrnaseq(sce_path)
}
\arguments{
\item{sce_path}{Input uploaded path}
}
\description{
Read the input single-cell RNA-seq dataset
from compressed RDS file. Accepts either:
(1) SingleCellExperiment
(2) Seurat object (converted to SingleCellExperiment)
}
