% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_cell_types_by_min_counts}
\alias{remove_cell_types_by_min_counts}
\title{Create a vector of metadata values that pass a minimum count}
\usage{
remove_cell_types_by_min_counts(
  grouped_frame,
  sce,
  metadata_column,
  min_counts = 2
)
}
\arguments{
\item{grouped_frame}{a data frame generated using \code{create_table_of_hetero_cat}}

\item{sce}{A SingleCellExperiment object}

\item{metadata_column}{The string name of a metadata column held within sce. Should be the same as the column used in grouped_frame.}

\item{min_counts}{the minimum counts to retain a category in grouped frame. Suggested minimum is 2.}
}
\description{
Create a vector of metadata values that pass a minimum count
}
