% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{suggest_genes_to_remove}
\alias{suggest_genes_to_remove}
\title{Suggest a set of redundant genes to remove}
\usage{
suggest_genes_to_remove(cmat, n_genes = 10)
}
\arguments{
\item{cmat}{A correlation matrix of \emph{the full expression} calculated as
\code{expression <- as.matrix(assay(sce, pref_assay)[markers$top_markers,])}
\code{cmat <- cor(t(expression))}}

\item{n_genes}{Number of genes to suggest to remove}
}
\description{
Suggest a set of redundant genes to remove
}
